# build/swagger.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: E - Commerce
  description: E - Commerce Server
  license:
    name: seongikCho
servers:
  - url: 'http://localhost:4444/'
components:
  parameters:
    productId:
      name: id
      in: path
      description: 제품 번호
      required: true
      schema:
        type: string
...
  securitySchemes:
    jwt:
      type: apiKey
      in: header
      name: Authorization
      description: JWT 토큰
  schemas:
    products:
      properties:
        name:
          type: string
          description: 상품 이름
          example: MagSafe형 iPhone 14 실리콘 케이스 - 카나리 옐로
        price:
          type: number
          description: 상품 가격
          example: 69000
        desc:
          type: string
          description: 상품에 대한 상세 정보
          example: Apple이 디자인한 이 MagSafe형 실리콘 케이스는 iPhone 14에 멋을 한층 더하는 동시에 폰을 든든하게 보호해줍니다.
    replies:
      properties:
        memo:
          type: string
          description: 제품에 대한 후기
          example: 저렴하게 잘 샀습니다. 다음에 또 주문할게요.
...
    Error:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
  examples:
    getAllProducts:
      value:
        message: successfully get all products from DB
        products:
          - _id: 6492e367a5ec58c9d161e1db
            name: 45mm 브라이트 오렌지 스포츠 밴드
            price: 65000
            desc: 스포츠 밴드는 특수 제작한 고탄성 불화탄성중합체로 만들어 튼튼하고 내구성이 뛰어나면서도 놀라우리만치 부드럽습니다. 부드러운 고밀도 소재의 밴드가 우아하게 손목을 감싸주어 피부에 닿는 느낌이 편안합니다. 또한 혁신적인 핀앤턱 잠금장치는 깔끔하게 떨어지는 착용감을 선사합니다.
            createdAt: '2023-06-21T11:47:51.230Z'
            updatedAt: '2023-06-21T11:47:51.230Z'
            __v: 0
          - _id: 649a91c8b20061abaf9e7df6
            name: iPad Pro 11
            price: 1249000
            desc: |-
              프로급 워크플로를 위한 막강한 성능과 온종일 가는 배터리 사용 시간을 구현하는 Apple M2 칩¹
              ProMotion, True Tone, P3의 넓은 색영역을 갖춘 27.9cm Liquid Retina 디스플레이각주²
              LiDAR 스캐너를 탑재한 프로급 카메라, 센터 스테이지 기술이 적용된 울트라 와이드 전면 카메라
              가장 빠른 Wi-Fi 연결성을 선사하는 Wi-Fi 6E 그리고 초고속 다운로드 및 고화질 스트리밍이 가능한 5G.³
              Apple Pencil(2세대),Magic Keyboard, Smart Keyboard Folio과 호환각주⁴
            createdAt: '2023-06-27T07:37:44.464Z'
            updatedAt: '2023-06-27T07:37:44.464Z'
            __v: 0
    getProduct:
      value:
        message: successfully get product from DB
        product:
          _id: 645c6e83b9e1ccb37c920ae7
          name: iPad Pro 11
          price: 1390000
          desc: |-
            프로급 워크플로를 위한 막강한 성능과 온종일 가는 배터리 사용 시간을 구현하는 Apple M2 칩¹
            ProMotion, True Tone, P3의 넓은 색영역을 갖춘 27.9cm Liquid Retina 디스플레이각주²
            LiDAR 스캐너를 탑재한 프로급 카메라, 센터 스테이지 기술이 적용된 울트라 와이드 전면 카메라
            가장 빠른 Wi-Fi 연결성을 선사하는 Wi-Fi 6E 그리고 초고속 다운로드 및 고화질 스트리밍이 가능한 5G.³
            Apple Pencil(2세대),Magic Keyboard, Smart Keyboard Folio과 호환각주⁴
          createdAt: '2023-05-11T04:26:44.005Z'
          updatedAt: '2023-05-11T04:26:44.005Z'
          __v: 0
...
    updateProduct:
      value:
        msg: successfully updated product by 649a91c8b20061abaf9e7df6
        product:
          _id: 649a91c8b20061abaf9e7df6
          name: update test
          price: 1234556
          desc: test1111111111111111111111111111
          createdAt: '2023-06-27T07:37:44.464Z'
          updatedAt: '2023-07-03T03:44:19.391Z'
          __v: 0
    deleteProduct:
      value:
        message: successfully deleted data by 645dea5be6b8f6d7c770f3bf
    createReplyToProduct:
      value:
        message: successfully created new reply
        reply:
          user: 64429ec62ea5616cfe5fe437
          memo: 저렴하게 잘 샀습니다. 다음에 또 주문할게요.
          reply_id: 645df643278345f7b7f21e6d
...
  responses:
...
    NoContent:
      description: '클라이언트의 요청이 성공적으로 처리되었지만, 응답 본문에 내용이 없음을 나타내는 상태 코드입니다.'
      content:
        application/json:
          schema:
            type: object
            example:
              status: 204
              success: true
              message: No content
    BadRequest:
      description: 클라이언트가 서버에 잘못된 요청을 보낸 경우 발생하는 오류입니다.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            status: 400
            success: false
            message: Bad Request
...
    NotFound:
      description: 클라이언트가 요청한 리소스를 서버에서 찾을 수 없을 때 발생하는 오류입니다.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            status: 404
            success: false
            message: Not Found
    InternalServerError:
      description: 서버 내부에서 예기치 않은 오류가 발생하여 클라이언트의 요청을 처리하지 못할 때 발생하는 상태 코드입니다.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            status: 500
            success: false
            message: Internal Server Error
paths:
  /products:
    get:
      tags:
        - Product
      summary: 모든 상품 조회
      consumes:
        - application/json
      produces:
        - application/json
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/products'
              examples:
                products:
                  $ref: '#/components/examples/getAllProducts'
        '204':
          $ref: '#/components/responses/NoContent'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - Product
      summary: 판매 상품 등록
      consumes:
        - application/json
      produces:
        - application/json
      security:
        - $ref: '#/components/securitySchemes/jwt'
      requestBody:
        x-name: body
        required: true
        content:
          x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/products'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/products'
              examples:
                productExample:
                  $ref: '#/components/examples/getProduct'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  '/products/{id}':
    get:
      tags:
        - Product
      summary: 제품의 상세 페이지 이동
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - $ref: '#/components/parameters/productId'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/products'
              examples:
                product:
                  $ref: '#/components/examples/getProduct'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags:
        - Product
      summary: 판매중인 상품 정보 수정
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - $ref: '#/components/parameters/productId'
      requestBody:
        x-name: body
        required: true
        content:
          x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/products'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/products'
              examples:
                product:
                  $ref: '#/components/examples/updateProduct'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags:
        - Product
      summary: 판매 상품 삭제
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - $ref: '#/components/parameters/productId'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/products'
              examples:
                deleteProducts:
                  $ref: '#/components/examples/deleteProduct'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - Product
      summary: 후기 작성 - 모든 사용자
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - $ref: '#/components/parameters/productId'
      requestBody:
        required: true
        content:
          x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/replies'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/replies'
              examples:
                deleteAllProducts:
                  $ref: '#/components/examples/createReplyToProduct'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
...